✨ add text message processing

+++ sender.py
@@ -1,3 +1,4 @@
+import re
 import logging
 
 from telegram import Bot, Update, Message, InlineKeyboardButton, InlineKeyboardMarkup
@@ -6,6 +7,7 @@ from settings import database
 
 
 logger = logging.getLogger(__name__)
+link = re.compile(r'https?://.\S+')
 
 
 def resend_message(bot: Bot, update: Update):
@@ -32,6 +34,7 @@ def resend_message(bot: Bot, update: Update):
                    {'document': message.document.file_id})
     elif message.text:
         logger.debug('Resending text...')
+        send_text_with_link(bot, message)
 
 
 def get_emoji_markup(emojis):
@@ -57,21 +60,44 @@ def emoji_callback(bot: Bot, update: Update):
                                       reply_markup=reply_markup)
 
 
-def send_media(message: Message, send_media_func, file_type_id: dict):
-
-    caption = f'by @{message.from_user.username}'
+def add_author(message: Message):
+    text = f'🌚 by @{message.from_user.username}'
     if message.forward_from and message.from_user.username != message.forward_from.username:
-        caption += f', from @{message.forward_from.username}'
+        if message.forward_from.username:
+            text += f', from @{message.forward_from.username}'
     if message.forward_from_chat and message.from_user.username != message.forward_from_chat.username:
-        caption += f', from @{message.forward_from_chat.username}'
+        if message.forward_from_chat.username:
+            text += f', from @{message.forward_from_chat.username}'
     if message.caption:
-        caption = message.caption + '\n' + caption  # fixme: caption size 0-200
+        text = message.caption + '\n' + text
+    if message.text:
+        text = message.text + '\n' + text
+    return text
 
-    init_emojis = {em: 0 for em in database.get_emojis(message.chat_id)}
-    reply_markup = get_emoji_markup(init_emojis)
+
+def send_media(message: Message, send_media_func, file_type_id: dict):
+    caption = add_author(message)
+
+    emojis = {em: 0 for em in database.get_emojis(message.chat_id)}
+    reply_markup = get_emoji_markup(emojis)
     sent_message = send_media_func(chat_id=message.chat_id,
-                                   caption=caption[:200],
+                                   caption=caption[:200],  # fixme: caption size 0-200
                                    reply_markup=reply_markup,
                                    **file_type_id)
+    database.add_message(sent_message, message.from_user, message.forward_from)
+    message.delete()
+
+
+def send_text_with_link(bot: Bot, message: Message):
+    if not link.findall(message.text):
+        return
+
+    text = add_author(message)
+
+    emojis = {em: 0 for em in database.get_emojis(message.chat_id)}
+    reply_markup = get_emoji_markup(emojis)
+    sent_message = bot.send_message(text=text,
+                                    chat_id=message.chat_id,
+                                    reply_markup=reply_markup)
     database.add_message(sent_message, message.from_user, message.forward_from)
     message.delete()
